Enabled alpha blending and changed pixelFormat to support it

+++ ArcadeApp/src/Shapes/AARectangle.cpp
@@ -59,8 +59,9 @@ bool AARectangle::ContainsPoint(const Vec2D& point) const
 
 AARectangle AARectangle::Inset(const AARectangle& rect, Vec2D& insets)
 {
-	return AARectangle(rect.GetTopLeftPoint() + insets, rect.GetWidth() - 2*insets.GetX(),
-				rect.GetHeight() - 2*insets.GetY());
+	return AARectangle(rect.GetTopLeftPoint() + insets, 
+				int(rect.GetWidth() - 2*insets.GetX()),
+				int(rect.GetHeight() - 2*insets.GetY()));
 }
 
 std::vector<Vec2D> AARectangle::GetPoints() const
